Split OBD connection observation from control
