+++ apps/server/vibesensor/adapters/obd/connection_runtime.py
@@ -18,6 +18,9 @@ from vibesensor.adapters.obd.connection_plan import (
 )
 from vibesensor.adapters.obd.elm327 import Elm327Session
 from vibesensor.adapters.obd.runtime_connection_control import ObdRuntimeConnectionControl
+from vibesensor.adapters.obd.runtime_connection_observation import (
+    ObdRuntimeConnectionObservation,
+)
 
 __all__ = ["ObdConnectionRuntime"]
 
@@ -31,7 +34,7 @@ class ObdConnectionRuntime:
     """Own session lifecycle, reconnect behavior, and blocking poll execution."""
 
     __slots__ = (
-        "_connection_control",
+        "_connection_observation",
         "_executor",
     )
 
@@ -39,13 +42,15 @@ class ObdConnectionRuntime:
         self,
         *,
         admin_client: ObdAdminClient,
+        connection_observation: ObdRuntimeConnectionObservation,
         connection_control: ObdRuntimeConnectionControl,
         session_factory: SessionFactory,
         monotonic: MonotonicFn = time.monotonic,
     ) -> None:
-        self._connection_control = connection_control
+        self._connection_observation = connection_observation
         self._executor = ObdConnectionExecutor(
             admin_client=admin_client,
+            connection_observation=connection_observation,
             connection_control=connection_control,
             session_factory=session_factory,
             monotonic=monotonic,
@@ -77,7 +82,7 @@ class ObdConnectionRuntime:
             selected_source,
             configured_mac,
             configured_name,
-        ) = self._connection_control.configured_device_snapshot()
+        ) = self._connection_observation.configured_device_snapshot()
         return plan_connection_step(
             ObdConnectionLoopSnapshot(
                 selected_source=selected_source,
@@ -86,7 +91,7 @@ class ObdConnectionRuntime:
                 has_session=session is not None,
                 session_device_mac=session_device_mac,
                 poll_wait_s=(
-                    self._connection_control.next_wait_s() if session is not None else None
+                    self._connection_observation.next_wait_s() if session is not None else None
                 ),
             ),
             idle_poll_s=_IDLE_POLL_S,
